+++ frontend/components/user/AuthButton.tsx
@@ -11,7 +11,7 @@ export default function AuthButton({ label, type = 'button', onClick }: Props) {
     <button
       type={type}
       onClick={onClick}
-      className="w-full text-white font-bold py-4 rounded-lg bg-gradient-to-r/ from-red-500 to-pink-600 hover:opacity-90 transition duration-150"
+      className="w-full text-white font-bold py-4 rounded-lg bg-linear-to-r from-red-500 to-pink-600 hover:opacity-90 transition duration-150"
     >
       {label}
     </button>
