+++ frontend/components/user/PasswordInput.tsx
@@ -14,7 +14,7 @@ export default function PasswordInput({ label, value, onChange }: Props) {
 
   return (
     <div className="space-y-2">
-      <label className="text-gray-700">{label}</label>
+      <label className="text-gray-700 block mb-1">{label}</label>
       <div className="relative">
         <Lock className="absolute left-3 top-1/2 -translate-y-1/2 h-5 w-5 text-gray-400" />
